Добавлены все методы для получения каталогов

+++ Backend/OldStore.API/Services/ICatalogsService.cs
@@ -6,6 +6,6 @@ namespace OldStore.API.Services
     {
         public Task<List<Catalog>> GetListCatalogsAsync();
 
-        public Task<Catalog> GetCatalogAsync(long id);
+        public Task<Catalog> GetCatalogAsync(int id);
     }
 }
